remove go.work; consolidate code
