--- go.work
@@ -1,6 +1,0 @@
-go 1.26.4
-
-use (
-	./ibc-link
-	./packages/go
-)
