+++ ibc-link/cmd/ibc/config.go
@@ -7,7 +7,7 @@ import (
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 
-	"github.com/cosmos/ibc/ibc-link/config"
+	"github.com/cosmos/ibc/ibc-link/internal/config"
 )
 
 // set in init()
