+++ ibc-link/go.mod
@@ -2,12 +2,21 @@ module github.com/cosmos/ibc/ibc-link
 
 go 1.26.4
 
-require github.com/spf13/cobra v1.10.2
+require (
+	github.com/goccy/go-yaml v1.19.2
+	github.com/pkg/errors v0.9.1
+	github.com/rs/zerolog v1.35.1
+	github.com/spf13/cobra v1.10.2
+	github.com/stretchr/testify v1.11.1
+)
 
 require (
-	github.com/goccy/go-yaml v1.19.2 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/inconshreveable/mousetrap v1.1.0 // indirect
-	github.com/pkg/errors v0.9.1 // indirect
+	github.com/mattn/go-colorable v0.1.14 // indirect
+	github.com/mattn/go-isatty v0.0.20 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
 	github.com/spf13/pflag v1.0.10 // indirect
-	github.com/stretchr/testify v1.11.1 // indirect
+	golang.org/x/sys v0.29.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
