+++ ibc-link/internal/logging/logging.go
@@ -1,3 +1,4 @@
+// Package logging provides logging functionality.
 package logging
 
 import (
@@ -6,6 +7,7 @@ import (
 	"github.com/rs/zerolog"
 )
 
+// DefaultLogger returns a default logger with INFO level.
 func DefaultLogger() zerolog.Logger {
 	return zerolog.New(os.Stdout).
 		Level(zerolog.InfoLevel).
@@ -14,6 +16,7 @@ func DefaultLogger() zerolog.Logger {
 		Logger()
 }
 
+// WithComponent returns a logger with the given component name.
 func WithComponent(logger zerolog.Logger, component string) zerolog.Logger {
 	return logger.With().Str("component", component).Logger()
 }
