--- packages/go/go.mod
@@ -1,11 +1,0 @@
-module github.com/cosmos/ibc/ibc-link/packages/go
-
-go 1.26.4
-
-require github.com/rs/zerolog v1.35.1
-
-require (
-	github.com/mattn/go-colorable v0.1.15 // indirect
-	github.com/mattn/go-isatty v0.0.22 // indirect
-	golang.org/x/sys v0.46.0 // indirect
-)
